Moves createCurrentData to CityCurrents model
